back-end/requirements.txt
# Core dependencies
fastapi
uvicorn
sqlmodel
apscheduler
httpx
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
python-multipart

# Development dependencies 
pytest
flake8
black
mypy
isort
